The module test code has been cleared.
